Improved: save action for selectProductStoreModal to be performed from facilityDetails page(#10)
